Adjusted Corn a bit, and added clearer commentation.
Corn needed to care about if its being planted in summer or not, since it can keep regrowing through fall. So I added a check for that in the code.

Also, I wrote some more commentation to help potential devs/users who read the code to understand what I'm doing and why.