Modified the program to include Sweet Gem Berry.
But it should be noted that Sweet Gems can't compete with the more usual crops for profit, because you can buy so many more of the normal crops than you can buy sweet gems with the same money.

I only included them to demonstrate that they're not proitable when being bought from the merchant.

They're considerably more profitable when obtained using the seed maker, especially when using fertilizers, but the effort required to get the necessary amount of rare seeds this way doesn't make them very effective in the first two years of the game.